requirements_dev.txt
flake8==7.2.0
ipdb==0.13.13
django-extensions==4.1